requirements.txt
# Core
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# OCR & Image Processing
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0
numpy>=1.24.0

# LLM Clients
openai>=1.0.0
anthropic>=0.5.0
tenacity>=8.2.0  # Retry logic

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
pytest-cov>=4.1.0
